requirements.txt
streamlit>=1.30.0
pyyaml>=6.0
pandas>=2.2.0
numpy>=1.25.0
matplotlib>=3.7.0
numpy-financial>=1.0.0
plotly